pyproject.toml
[project]
name = "terraform-google-networking"
version = "0.1.1"
requires-python = ">=3.11"
dependencies = [
    "Quart>=0.19.0",
    "cryptography>=44.0.1",
    "gcloud-aio-auth>=5.4.2",
    "gcloud-aio-storage>=9.5.0",
    "GitPython>=3.1.44",
    "google-api-core>=2.25.1",
    "google-auth>=2.35.0",
    "Hypercorn>=0.18.0",
    "PyYAML>=6.0.2",
    "tomli>=2.0.1",
]

[tool.setuptools]
packages = []

